requirements.txt
pandas>=1.5.0
pyyaml>=6.0
ollama>=0.1.0
pytest>=7.0.0